backend/requirements.txt
# Web框架
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# 数据库
sqlalchemy==2.0.36
aiomysql==0.2.0
alembic==1.14.0
redis==5.2.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# LangChain和向量数据库
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
chromadb==0.5.0

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
paddleocr==2.8.0
paddlepaddle==3.0.0

# 语音识别 - 暂时注释,后续单独安装
# openai-whisper==20231117

# LLM客户端
openai==1.50.0
ollama==0.3.0

# 工具库
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
aiofiles==24.1.0
pillow==10.4.0
pymysql==1.1.0

